Move init command into a separate internal package.
