+++ cmd/conduit/internal/init/conduit.yml.example
